reverting a change that seems to have broken IPv6.. I feel like this will bite me again but for now it works better without the int stuff.. was tested on 16.04

+++ www/include/functions_general.inc.php
@@ -635,7 +635,7 @@ function ip_mangle_gmp($ip="", $format="default") {
 
     // If we get here, then the input must be in numeric format (1)
     else {
-        $ip = gmp_init((int) strval($ip), 10);
+        $ip = gmp_init(strval($ip), 10);
         if ($format == "default") {
             if(is_ipv4($ip))
                 $format = "dotted";
